+++ CHANGELOG.md
@@ -1,9 +1,12 @@
 Changelog
 =========
 
-## TBD
+## 6.31.0 (2025-01-13)
 
 ### Enhancements
+
+* Disabled assertions on release build.
+  [1717](https://github.com/bugsnag/bugsnag-cocoa/pull/1717)
 
 * Optimized feature flags store.
   [1715](https://github.com/bugsnag/bugsnag-cocoa/pull/1715)